refactor[OK-143] Add user-boards directory for user boards related components
