+++ client/src/app/auth/signup/email/page.tsx
@@ -1,6 +1,6 @@
 "use client";
-import AuthHeaderButtons from "@/components/AuthHeaderButtons";
-import ExpandedLogoAbbreviation from "@/components/ExpandedLogoAbbreviation";
+import AuthHeaderButtons from "@/components/user-boards/AuthHeaderButtons";
+import ExpandedLogoAbbreviation from "@/components/user-boards/ExpandedLogoAbbreviation";
 import EmailSignupCard from "./EmailSignupCard";
 
 const SignupWithEmail: React.FC = () => {
