+++ client/src/app/user-boards/page.tsx
@@ -1,9 +1,9 @@
 "use client";
-import Sidebar from "@/components/Sidebar";
-import Header from "@/components/Header";
-import BoardTable from "@/components/BoardTable";
-import UserBoardsActionButtons from "@/components/UserBoardsActionButtons";
-import TabButtons from "@/components/TabButtons";
+import Sidebar from "@/components/user-boards/Sidebar";
+import Header from "@/components/user-boards/Header";
+import BoardTable from "@/components/user-boards/BoardTable";
+import UserBoardsActionButtons from "@/components/user-boards/UserBoardsActionButtons";
+import TabButtons from "@/components/user-boards/TabButtons";
 
 export default function UserBoards() {
   const tabs = [{ label: "All" }, { label: "Shared by others" }, { label: "Archived" }];
