+++ client/src/components/ui/form.tsx
@@ -78,7 +78,7 @@ const FormLabel = React.forwardRef<
   React.ElementRef<typeof LabelPrimitive.Root>,
   React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root>
 >(({ className, ...props }, ref) => {
-  const { error, formItemId } = useFormField();
+  const { formItemId } = useFormField();
 
   return <Label ref={ref} className={className} htmlFor={formItemId} {...props} />;
 });
