+++ client/src/components/user-boards/AuthHeaderButtons.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import ThemeToggle from "./ThemeToggle";
 
 const AuthHeaderButtons: React.FC = () => {
