+++ client/src/components/user-boards/ExpandedLogoAbbreviation.tsx
