+++ client/src/components/user-boards/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 import { FC, useEffect, useState } from "react";
 import { useTheme } from "next-themes";
 import { Moon, Sun } from "lucide-react";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 
 const ThemeToggle: FC = () => {
   const [mounted, setMounted] = useState(false);
